feat: more type structure simplifications
